# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AI Dianping Agent Platform API
  version: 1.2.0
servers:
  - url: http://localhost:8081
security:
  - bearerAuth: []
paths:
...
  /api/v1/agents/{id}:
    parameters: [ { $ref: '#/components/parameters/IdPath' } ]
    get: { summary: Get agent, responses: { '200': { $ref: '#/components/responses/Success' } } }
...
  /api/v1/agent-runs:
    get:
      summary: List agent run history
      description: >
        Ordinary AGENT_RUN users only see their own runs.
        RUN_INSPECT or ADMIN can inspect all runs inside the current scope.
      parameters:
        - { $ref: '#/components/parameters/TenantHeader' }
        - { $ref: '#/components/parameters/WorkspaceHeader' }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: size, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
        - { name: agentId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/RunStatus' } }
        - { name: createdFrom, in: query, schema: { type: string, format: date-time } }
        - { name: createdTo, in: query, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Run history page
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PageResponse' }
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/AgentRunSummary' }
    post:
      summary: Create an Agent run
      parameters:
        - { $ref: '#/components/parameters/TenantHeader' }
        - { $ref: '#/components/parameters/WorkspaceHeader' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentRunRequest' }
      responses:
        '200':
          description: Run queued
          content: { application/json: { schema: { $ref: '#/components/schemas/RunCreated' } } }
  /api/v1/agent-runs/{runId}:
    parameters:
      - { $ref: '#/components/parameters/RunIdPath' }
      - { $ref: '#/components/parameters/TenantHeader' }
      - { $ref: '#/components/parameters/WorkspaceHeader' }
    get:
      summary: Get run details
      responses:
        '200':
          description: Run detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentRunDetail' }
  /api/v1/agent-runs/{runId}/events:
    parameters:
      - { $ref: '#/components/parameters/RunIdPath' }
      - { $ref: '#/components/parameters/TenantHeader' }
      - { $ref: '#/components/parameters/WorkspaceHeader' }
    get:
      summary: Stream run lifecycle events
      parameters:
        - { name: Last-Event-ID, in: header, required: false, schema: { type: integer, format: int64, minimum: 0 } }
      responses:
        '200':
          description: SSE stream of lifecycle events (not token streaming)
          content:
            text/event-stream:
              schema: { $ref: '#/components/schemas/AgentRunEvent' }
  /api/v1/agent-runs/{runId}/cancel:
    parameters: [ { $ref: '#/components/parameters/RunIdPath' } ]
    post: { summary: Cancel run, responses: { '200': { content: { application/json: { schema: { $ref: '#/components/schemas/RunCreated' } } }, description: Run cancelled } } }
  /api/v1/agent-runs/{runId}/retry:
    parameters: [ { $ref: '#/components/parameters/RunIdPath' } ]
    post: { summary: Retry run, responses: { '200': { content: { application/json: { schema: { $ref: '#/components/schemas/RunCreated' } } }, description: Run queued } } }
  /api/v1/agent-runs/{runId}/resume:
    parameters: [ { $ref: '#/components/parameters/RunIdPath' } ]
    post: { summary: Resume run, responses: { '200': { content: { application/json: { schema: { $ref: '#/components/schemas/RunCreated' } } }, description: Run queued } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    TenantHeader: { name: X-Tenant-Id, in: header, required: true, schema: { type: string, maxLength: 64 } }
    WorkspaceHeader: { name: X-Workspace-Id, in: header, required: true, schema: { type: string, maxLength: 64 } }
    IdPath: { name: id, in: path, required: true, schema: { type: string, maxLength: 64 } }
    RunIdPath: { name: runId, in: path, required: true, schema: { type: string, maxLength: 64 } }
...
  responses:
    Success:
      description: Successful response
      content: { application/json: { schema: { type: object } } }
  schemas:
...
    PageResponse:
      type: object
      required: [items, total, page, size]
      properties:
        items: { type: array, items: { type: object } }
        total: { type: integer, format: int64, minimum: 0 }
        page: { type: integer, minimum: 1 }
        size: { type: integer, minimum: 1 }
...
    RunStatus:
      type: string
      enum: [CREATED, QUEUED, RUNNING, WAITING_FOR_USER, WAITING_FOR_APPROVAL, COMPLETED, FAILED, CANCELLED, TIMED_OUT]
    AgentRunSummary:
      type: object
      required: [runId, agentId, agentVersion, status, sessionId]
      properties:
        runId: { type: string }
        agentId: { type: string }
        agentVersion: { type: integer, minimum: 1 }
        status: { $ref: '#/components/schemas/RunStatus' }
        sessionId: { type: string }
        errorCode: { type: string, nullable: true }
        errorMessage: { type: string, nullable: true }
        queuedAt: { type: string, format: date-time, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        finishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
    AgentRunDetail:
      allOf:
        - { $ref: '#/components/schemas/AgentRunSummary' }
        - type: object
          properties:
            conversationId: { type: string }
            traceId: { type: string }
            versionSnapshot: { type: object }
            output: { type: object, nullable: true }
            deadlineAt: { type: string, format: date-time, nullable: true }
    AgentRunEvent:
      type: object
      required: [sequence, runId, type, payload, createdAt]
      properties:
        sequence: { type: integer, format: int64 }
        runId: { type: string }
        type: { type: string }
        payload: { type: object }
        createdAt: { type: string, format: date-time }
    AgentRunRequest:
      type: object
      additionalProperties: false
      required: [agentId, agentVersion, sessionId, input, responseMode]
      properties:
        agentId: { type: string, maxLength: 64 }
        agentVersion: { type: integer, minimum: 1 }
        sessionId: { type: string, maxLength: 128 }
        input:
          type: object
          required: [text]
          properties:
            text: { type: string, maxLength: 20000 }
            parts: { type: array, items: { type: object } }
            attachments: { type: array, items: { type: object } }
            referenceUris: { type: array, items: { type: string } }
        responseMode: { type: string, enum: [BLOCKING, STREAM] }
        metadata: { type: object }
    RunCreated:
      type: object
      additionalProperties: false
      required: [runId, status, agentDefinitionId, agentCode, agentVersion]
      properties:
        runId: { type: string }
        status: { $ref: '#/components/schemas/RunStatus' }
        agentDefinitionId: { type: string }
        agentCode: { type: string }
        agentVersion: { type: integer, minimum: 1 }